+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 Features:
 
+  - [lsp] Code action to complete constructor with _public_ properties
   - [php] Bump min. PHP version to 7.4
   - [php] Fix PHP 8.1 deprecations
   - [config] JSON schema support